Create Test\Cases namespace and override Admin tests from AdminCouponTestCases and move the ADDR to that class
